refactor(structex): split tests/conftest.py into tests/fixtures plugins (VAL-P1-007)
Extract all 46 fixtures from the 3,044-line tests/conftest.py into eight
cohesive tests/fixtures/*.py plugin modules loaded via pytest_plugins;
conftest drops to 561 lines and retains only bootstrap, optional-dependency
skip markers, marker registration, and skip-count reporting. Fixture bodies
are byte-identical verbatim slices; every autouse fixture lives in
tests/fixtures/autouse.py in its original definition order so autouse
setup/teardown ordering is unchanged (verified via pytest --setup-plan diff).
Only skip markers were ever imported from tests.conftest and they remain
there, so import paths stay stable for every test.
